chore(headless): fix TS errors
