+++ packages/headless-styles/tests/tag/tagCSS.test.ts
@@ -1,4 +1,4 @@
-import { getTagProps, getTagWithIconProps } from '../../src'
+import { getTagProps } from '../../src'
 
 describe('Tag CSS', () => {
   const baseClass = 'ps-tag'
